v1/requirements.txt
customtkinter==5.2.2
openpyxl==3.1.2
mysql-connector-python==8.0.33